Move weather logic to weather.js file
